Move files to correct places and adjust paths
