+++ C++/CMakeLists.txt
@@ -143,7 +143,7 @@ target_link_libraries(singly_linked_list test_runner)
 
 # Doubly-linked list
 add_executable(doubly_linked_list
-        test/data_structures/doubly_linked_list.cpp)
+        test/data_structure/linked_list/doubly_linked_list.cpp)
 target_link_libraries(doubly_linked_list test_runner)
 
 # ---
